requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2==2.8.4
SQLAlchemy==1.3.10
Werkzeug==0.16.0
coverage==4.5.4
GitPython==3.0.4
pre-commit==1.20.0
pytest==5.2.2
pytest-cov==2.8.1
python-dotenv==0.10.3
Flask-Migrate==2.5.2
